requirements.txt
streamlit
python-dotenv
pypdf
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-google-genai
faiss-cpu
sentence-transformers
google-generativeai
